requirements.txt
aiohttp~=3.7.4
aiofile~=3.7.2
PyYAML~=6.0